Fixed plug-in build problem
The problem happened because of the latest update to Selenium. We transitively depend on Selenium (via Crawljax), but Crawljax depends on the latest version of Selenium. Because (for the moment) we are not generating the .classpath file for Eclipse, the versions of all dependencies are hard-coded in this file. Whenever a new version of Selenium comes out, the latest version is fetched (when building), and the .jar file names will be different. Consequently, the update to Selenium would break the building of the Eclipse plugin.